js: CFX_beautiful -> CFX
